==> Codes/5.TopicModelingLDA_2.1.ipynb <==
###' ################################################################################
###'
###' IMPORT LIBRARIES
###'
###'

import numpy as np
import pandas as pd
from pathlib import Path
import random
import torch
from keybert import KeyBERT

import ast
import umap.umap_ as umap
import matplotlib.pyplot as plt
import matplotlib.colors as mcolors
from sklearn.cluster import KMeans
from sklearn.cluster import SpectralClustering
from sklearn.cluster import AgglomerativeClustering
from sklearn.preprocessing import StandardScaler
from mpl_toolkits.mplot3d import Axes3D
from sklearn.manifold import SpectralEmbedding

### LDA
import gensim
import gensim.corpora as corpora
from sklearn.feature_extraction.text import TfidfVectorizer
from gensim.models import Phrases

### BERT
from bertopic import BERTopic
from sklearn.feature_extraction.text import CountVectorizer
from bertopic.vectorizers import ClassTfidfTransformer

# Load dataset
data_dir = Path(r"C:\Users\Hyemi\Python\TopicModeling\Data")
data_dir.mkdir(parents=True, exist_ok=True) 

embedding_files = [
    "articles_4_clustering1.csv",
    "articles_4_clustering2.csv",
    "articles_4_clustering3.csv",
    "articles_4_clustering4.csv"
]

embedding_file_paths = [data_dir / file for file in embedding_files]


embedding_dfs = []
for file_path in embedding_file_paths:
    if file_path.exists():  # Check if the file exists before reading
        embedding_dfs.append(pd.read_csv(file_path))

df1a = embedding_dfs[0]
df2a = embedding_dfs[1]
df3a = embedding_dfs[2]
df4a = embedding_dfs[3]

df3a

## Group 1

df1 = df1a
print(df1['Cluster3'].value_counts())
print(df1['Cluster4'].value_counts())
print(df1['Cluster5'].value_counts())

###########  BEST so far

np.random.seed(40)
random.seed(40)
torch.manual_seed(40) 

abstracts = df1['Abstract_join'].dropna().tolist()
cluster_labels = df1['Cluster3'].astype(int).tolist() 

# Apply BERTopic with Predefined Cluster Labels
vectorizer_model = CountVectorizer(ngram_range=(2, 4), stop_words='english')
bertopic_model = BERTopic(vectorizer_model=vectorizer_model, ctfidf_model=ClassTfidfTransformer())

# Fit BERTopic using predefined cluster labels
topics, _ = bertopic_model.fit_transform(abstracts, y=cluster_labels)  
bertopic_model.reduce_topics(abstracts, nr_topics=4)  

# Store topics in dataframe
#df['topic'] = topics

# Get Top Words per Cluster
topic_summary = bertopic_model.get_topics()

# Print Cluster Topics
for cluster in sorted(df1['Cluster3'].unique()):
    if cluster in topic_summary:  # Ensure the cluster exists in the BERTopic output
        print(f"Cluster {cluster}: {', '.join([word[0] for word in topic_summary[cluster][:10]])}")

###########  BEST so far

np.random.seed(42)
random.seed(42)
torch.manual_seed(42) 

abstracts = df1['Abstract_join'].dropna().tolist()
cluster_labels = df1['Cluster4'].astype(int).tolist() 

# Apply BERTopic with Predefined Cluster Labels
vectorizer_model = CountVectorizer(ngram_range=(2, 4), stop_words='english')
bertopic_model = BERTopic(vectorizer_model=vectorizer_model, ctfidf_model=ClassTfidfTransformer())

# Fit BERTopic using predefined cluster labels
topics, _ = bertopic_model.fit_transform(abstracts, y=cluster_labels)  
bertopic_model.reduce_topics(abstracts, nr_topics=5)  

# Store topics in dataframe
#df['topic'] = topics

# Get Top Words per Cluster
topic_summary = bertopic_model.get_topics()

# Print Cluster Topics
for cluster in sorted(df1['Cluster4'].unique()):
    if cluster in topic_summary:  # Ensure the cluster exists in the BERTopic output
        print(f"Cluster {cluster}: {', '.join([word[0] for word in topic_summary[cluster][:10]])}")

###########  BEST so far

np.random.seed(42)
random.seed(42)
torch.manual_seed(42) 

abstracts = df1['Abstract_join'].dropna().tolist()
cluster_labels = df1['Cluster5'].astype(int).tolist() 

# Apply BERTopic with Predefined Cluster Labels
vectorizer_model = CountVectorizer(ngram_range=(2, 4), stop_words='english')
bertopic_model = BERTopic(vectorizer_model=vectorizer_model, ctfidf_model=ClassTfidfTransformer())

# Fit BERTopic using predefined cluster labels
topics, _ = bertopic_model.fit_transform(abstracts, y=cluster_labels)  
bertopic_model.reduce_topics(abstracts, nr_topics=6)  

# Store topics in dataframe
#df['topic'] = topics

# Get Top Words per Cluster
topic_summary = bertopic_model.get_topics()

# Print Cluster Topics
for cluster in sorted(df1['Cluster5'].unique()):
    if cluster in topic_summary:  # Ensure the cluster exists in the BERTopic output
        print(f"Cluster {cluster}: {', '.join([word[0] for word in topic_summary[cluster][:10]])}")

## Group 2

df2 = df2a
print(df2['Cluster3'].value_counts())
print(df2['Cluster4'].value_counts())
print(df2['Cluster5'].value_counts())

###########  BEST so far

np.random.seed(42)
random.seed(42)
torch.manual_seed(42) 

abstracts = df2['Abstract_join'].dropna().tolist()
cluster_labels = df2['Cluster3'].astype(int).tolist()  

# Apply BERTopic with Predefined Cluster Labels
vectorizer_model = CountVectorizer(ngram_range=(2, 4), stop_words='english')
bertopic_model = BERTopic(vectorizer_model=vectorizer_model, ctfidf_model=ClassTfidfTransformer())

# Fit BERTopic using predefined cluster labels
topics, _ = bertopic_model.fit_transform(abstracts, y=cluster_labels)  
bertopic_model.reduce_topics(abstracts, nr_topics=4)  

# Store topics in dataframe
#df['topic'] = topics

# Get Top Words per Cluster
topic_summary = bertopic_model.get_topics()

# Print Cluster Topics
for cluster in sorted(df2['Cluster3'].unique()):
    if cluster in topic_summary:  # Ensure the cluster exists in the BERTopic output
        print(f"Cluster {cluster}: {', '.join([word[0] for word in topic_summary[cluster][:10]])}")

###########  BEST so far

np.random.seed(42)
random.seed(42)
torch.manual_seed(42) 

abstracts = df2['Abstract_join'].dropna().tolist()
cluster_labels = df2['Cluster4'].astype(int).tolist()  

# Apply BERTopic with Predefined Cluster Labels
vectorizer_model = CountVectorizer(ngram_range=(2, 4), stop_words='english')
bertopic_model = BERTopic(vectorizer_model=vectorizer_model, ctfidf_model=ClassTfidfTransformer())

# Fit BERTopic using predefined cluster labels
topics, _ = bertopic_model.fit_transform(abstracts, y=cluster_labels)  
bertopic_model.reduce_topics(abstracts, nr_topics=5)  

# Store topics in dataframe
#df['topic'] = topics

# Get Top Words per Cluster
topic_summary = bertopic_model.get_topics()

# Print Cluster Topics
for cluster in sorted(df2['Cluster4'].unique()):
    if cluster in topic_summary:  # Ensure the cluster exists in the BERTopic output
        print(f"Cluster {cluster}: {', '.join([word[0] for word in topic_summary[cluster][:10]])}")

###########  BEST so far

np.random.seed(42)
random.seed(42)
torch.manual_seed(42) 

abstracts = df2['Abstract_join'].dropna().tolist()
cluster_labels = df2['Cluster5'].astype(int).tolist()  

# Apply BERTopic with Predefined Cluster Labels
vectorizer_model = CountVectorizer(ngram_range=(2, 4), stop_words='english')
bertopic_model = BERTopic(vectorizer_model=vectorizer_model, ctfidf_model=ClassTfidfTransformer())

# Fit BERTopic using predefined cluster labels
topics, _ = bertopic_model.fit_transform(abstracts, y=cluster_labels)  
bertopic_model.reduce_topics(abstracts, nr_topics=6)  

# Store topics in dataframe
#df['topic'] = topics

# Get Top Words per Cluster
topic_summary = bertopic_model.get_topics()

# Print Cluster Topics
for cluster in sorted(df2['Cluster5'].unique()):
    if cluster in topic_summary:  # Ensure the cluster exists in the BERTopic output
        print(f"Cluster {cluster}: {', '.join([word[0] for word in topic_summary[cluster][:10]])}")

## GROUP3

df3 = df3a
print(df3['Cluster4'].value_counts())
print(df3['Cluster5'].value_counts())
print(df3['Cluster8'].value_counts())
print(df3['Cluster9'].value_counts())

np.random.seed(40)
random.seed(40)
torch.manual_seed(40) 

abstracts = df3['Abstract_join'].dropna().tolist()
cluster_labels = df3['Cluster4'].astype(int).tolist()  

# Apply BERTopic with Predefined Cluster Labels
vectorizer_model = CountVectorizer(ngram_range=(2, 4), stop_words='english')
bertopic_model = BERTopic(vectorizer_model=vectorizer_model, ctfidf_model=ClassTfidfTransformer())

# Fit BERTopic using predefined cluster labels
topics, _ = bertopic_model.fit_transform(abstracts, y=cluster_labels)  
bertopic_model.reduce_topics(abstracts, nr_topics=5)  

# Store topics in dataframe
#df['topic'] = topics

# Get Top Words per Cluster
topic_summary = bertopic_model.get_topics()

# Print Cluster Topics
for cluster in sorted(df3['Cluster4'].unique()):
    if cluster in topic_summary:  # Ensure the cluster exists in the BERTopic output
        print(f"Cluster {cluster}: {', '.join([word[0] for word in topic_summary[cluster][:10]])}")

np.random.seed(40)
random.seed(40)
torch.manual_seed(40) 

abstracts = df3['Abstract_join'].dropna().tolist()
cluster_labels = df3['Cluster5'].astype(int).tolist()  

# Apply BERTopic with Predefined Cluster Labels
vectorizer_model = CountVectorizer(ngram_range=(2, 4), stop_words='english')
bertopic_model = BERTopic(vectorizer_model=vectorizer_model, ctfidf_model=ClassTfidfTransformer())

# Fit BERTopic using predefined cluster labels
topics, _ = bertopic_model.fit_transform(abstracts, y=cluster_labels)  
bertopic_model.reduce_topics(abstracts, nr_topics=6)  

# Store topics in dataframe
#df['topic'] = topics

# Get Top Words per Cluster
topic_summary = bertopic_model.get_topics()

# Print Cluster Topics
for cluster in sorted(df3['Cluster5'].unique()):
    if cluster in topic_summary:  # Ensure the cluster exists in the BERTopic output
        print(f"Cluster {cluster}: {', '.join([word[0] for word in topic_summary[cluster][:10]])}")

np.random.seed(40)
random.seed(40)
torch.manual_seed(40) 

abstracts = df3['Abstract_join'].dropna().tolist()
cluster_labels = df3['Cluster8'].astype(int).tolist()  

# Apply BERTopic with Predefined Cluster Labels
vectorizer_model = CountVectorizer(ngram_range=(2, 4), stop_words='english')
bertopic_model = BERTopic(vectorizer_model=vectorizer_model, ctfidf_model=ClassTfidfTransformer())

# Fit BERTopic using predefined cluster labels
topics, _ = bertopic_model.fit_transform(abstracts, y=cluster_labels)  
bertopic_model.reduce_topics(abstracts, nr_topics=9)  

# Store topics in dataframe
#df['topic'] = topics

# Get Top Words per Cluster
topic_summary = bertopic_model.get_topics()

# Print Cluster Topics
for cluster in sorted(df3['Cluster8'].unique()):
    if cluster in topic_summary:  # Ensure the cluster exists in the BERTopic output
        print(f"Cluster {cluster}: {', '.join([word[0] for word in topic_summary[cluster][:10]])}")

np.random.seed(40)
random.seed(40)
torch.manual_seed(40) 

abstracts = df3['Abstract_join'].dropna().tolist()
cluster_labels = df3['Cluster9'].astype(int).tolist()  

# Apply BERTopic with Predefined Cluster Labels
vectorizer_model = CountVectorizer(ngram_range=(2, 4), stop_words='english')
bertopic_model = BERTopic(vectorizer_model=vectorizer_model, ctfidf_model=ClassTfidfTransformer())

# Fit BERTopic using predefined cluster labels
topics, _ = bertopic_model.fit_transform(abstracts, y=cluster_labels)  
bertopic_model.reduce_topics(abstracts, nr_topics=10)  

# Store topics in dataframe
#df['topic'] = topics

# Get Top Words per Cluster
topic_summary = bertopic_model.get_topics()

# Print Cluster Topics
for cluster in sorted(df3['Cluster9'].unique()):
    if cluster in topic_summary:  # Ensure the cluster exists in the BERTopic output
        print(f"Cluster {cluster}: {', '.join([word[0] for word in topic_summary[cluster][:10]])}")

## GROUP4

df4 = df4a
print(df4['Cluster4'].value_counts())
print(df4['Cluster5'].value_counts())
print(df4['Cluster9'].value_counts())
print(df4['Cluster10'].value_counts())

###########  BEST so far

np.random.seed(40)
random.seed(40)
torch.manual_seed(40) 

abstracts = df4['Abstract_join'].dropna().tolist()
cluster_labels = df4['Cluster4'].astype(int).tolist()  

# Apply BERTopic with Predefined Cluster Labels
vectorizer_model = CountVectorizer(ngram_range=(2, 4), stop_words='english')
bertopic_model = BERTopic(vectorizer_model=vectorizer_model, ctfidf_model=ClassTfidfTransformer())

# Fit BERTopic using predefined cluster labels
topics, _ = bertopic_model.fit_transform(abstracts, y=cluster_labels)  
bertopic_model.reduce_topics(abstracts, nr_topics=5)  

# Store topics in dataframe
#df['topic'] = topics

# Get Top Words per Cluster
topic_summary = bertopic_model.get_topics()

# Print Cluster Topics
for cluster in sorted(df4['Cluster4'].unique()):
    if cluster in topic_summary:  # Ensure the cluster exists in the BERTopic output
        print(f"Cluster {cluster}: {', '.join([word[0] for word in topic_summary[cluster][:10]])}")

###########  BEST so far

np.random.seed(40)
random.seed(40)
torch.manual_seed(40) 

abstracts = df4['Abstract_join'].dropna().tolist()
cluster_labels = df4['Cluster5'].astype(int).tolist()  

# Apply BERTopic with Predefined Cluster Labels
vectorizer_model = CountVectorizer(ngram_range=(2, 4), stop_words='english')
bertopic_model = BERTopic(vectorizer_model=vectorizer_model, ctfidf_model=ClassTfidfTransformer())

# Fit BERTopic using predefined cluster labels
topics, _ = bertopic_model.fit_transform(abstracts, y=cluster_labels)  
bertopic_model.reduce_topics(abstracts, nr_topics=6)  

# Store topics in dataframe
#df['topic'] = topics

# Get Top Words per Cluster
topic_summary = bertopic_model.get_topics()

# Print Cluster Topics
for cluster in sorted(df4['Cluster5'].unique()):
    if cluster in topic_summary:  # Ensure the cluster exists in the BERTopic output
        print(f"Cluster {cluster}: {', '.join([word[0] for word in topic_summary[cluster][:10]])}")

###########  BEST so far

np.random.seed(40)
random.seed(40)
torch.manual_seed(40) 

abstracts = df4['Abstract_join'].dropna().tolist()
cluster_labels = df4['Cluster9'].astype(int).tolist()  

# Apply BERTopic with Predefined Cluster Labels
vectorizer_model = CountVectorizer(ngram_range=(2, 4), stop_words='english')
bertopic_model = BERTopic(vectorizer_model=vectorizer_model, ctfidf_model=ClassTfidfTransformer())

# Fit BERTopic using predefined cluster labels
topics, _ = bertopic_model.fit_transform(abstracts, y=cluster_labels)  
bertopic_model.reduce_topics(abstracts, nr_topics=10)  

# Store topics in dataframe
#df['topic'] = topics

# Get Top Words per Cluster
topic_summary = bertopic_model.get_topics()

# Print Cluster Topics
for cluster in sorted(df4['Cluster9'].unique()):
    if cluster in topic_summary:  # Ensure the cluster exists in the BERTopic output
        print(f"Cluster {cluster}: {', '.join([word[0] for word in topic_summary[cluster][:10]])}")

###########  BEST so far

np.random.seed(40)
random.seed(40)
torch.manual_seed(40) 

abstracts = df4['Abstract_join'].dropna().tolist()
cluster_labels = df4['Cluster10'].astype(int).tolist()  

# Apply BERTopic with Predefined Cluster Labels
vectorizer_model = CountVectorizer(ngram_range=(2, 4), stop_words='english')
bertopic_model = BERTopic(vectorizer_model=vectorizer_model, ctfidf_model=ClassTfidfTransformer())

# Fit BERTopic using predefined cluster labels
topics, _ = bertopic_model.fit_transform(abstracts, y=cluster_labels)  
bertopic_model.reduce_topics(abstracts, nr_topics=11)  

# Store topics in dataframe
#df['topic'] = topics

# Get Top Words per Cluster
topic_summary = bertopic_model.get_topics()

# Print Cluster Topics
for cluster in sorted(df4['Cluster10'].unique()):
    if cluster in topic_summary:  # Ensure the cluster exists in the BERTopic output
        print(f"Cluster {cluster}: {', '.join([word[0] for word in topic_summary[cluster][:10]])}")
